+++ app/src/androidTest/java/com/github/gogetters/letsgo/activities/ChatActivityTest.kt
@@ -1,9 +1,10 @@
-package com.github.gogetters.letsgo
+package com.github.gogetters.letsgo.activities
 
 import android.widget.EditText
 import android.widget.ListView
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.gogetters.letsgo.R
 import com.github.gogetters.letsgo.chat.ChatMessage
 import junit.framework.Assert.*
 import org.junit.Rule
@@ -20,14 +21,14 @@ class ChatActivityTest {
         val scenario = activityRule.scenario
         scenario.onActivity { activity ->
             val testText = "Hello"
-            val entryText: EditText = activity.findViewById(R.id.editText)
+            val entryText: EditText = activity.findViewById(R.id.chat_editText_input)
             entryText.setText(testText)
             activity.sendMessage(null) // argument is not used
 
             // Check if entry field was emptied
             assertTrue(entryText.text.toString().isEmpty())
             // Check if the message was added to the listview
-            val listView: ListView = activity.findViewById(R.id.messages_view)
+            val listView: ListView = activity.findViewById(R.id.chat_listView_messages)
             val item: ChatMessage = (listView.getItemAtPosition(0) as ChatMessage)
             assertEquals(item.getText(), testText)
         }
