+++ app/src/main/java/com/github/gogetters/letsgo/database/Database.kt
@@ -1,0 +1,4 @@
+package com.github.gogetters.letsgo.database
+
+class Database {
+}
