+++ app/src/main/res/layout/activity_chat.xml
@@ -6,10 +6,10 @@
     android:orientation="vertical"
     android:clipToPadding="false"
     android:focusableInTouchMode="true"
-    tools:context="com.github.gogetters.letsgo.ChatActivity">
+    tools:context="com.github.gogetters.letsgo.activities.ChatActivity">
 
     <ListView
-        android:id="@+id/messages_view"
+        android:id="@+id/chat_listView_messages"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="2"
@@ -22,7 +22,7 @@
         android:orientation="horizontal">
 
         <EditText
-            android:id="@+id/editText"
+            android:id="@+id/chat_editText_input"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="2"
