+++ app/src/main/res/layout/activity_game.xml
@@ -5,7 +5,7 @@
     android:id="@+id/gameConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameActivity">
+    tools:context=".activities.GameActivity">
 
     <FrameLayout
         android:id="@+id/gameBoardFrame"
