+++ app/src/main/res/layout/activity_profile.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Profile"
+    tools:context=".activities.ProfileActivity"
     android:background="@drawable/gradient"
     android:orientation="vertical">
