+++ app/src/main/res/layout/my_message.xml
@@ -10,7 +10,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/message_body"
+        android:id="@+id/chat_textView_messageBody"
         android:background="@drawable/my_message"
         android:textColor="#fff"
         android:padding="10dp"
