+++ app/src/main/res/layout/their_message.xml
@@ -8,7 +8,7 @@
     android:clipToPadding="false">
 
     <View
-        android:id="@+id/avatar"
+        android:id="@+id/chat_view_avatar"
         android:layout_alignParentLeft="true"
         android:scaleType="centerInside"
         android:background="@drawable/circle"
@@ -16,10 +16,10 @@
         android:layout_height="34dp" />
 
     <TextView
-        android:id="@+id/name"
+        android:id="@+id/chat_textView_name"
         android:layout_marginLeft="15dp"
-        android:layout_toRightOf="@+id/avatar"
-        android:layout_alignTop="@+id/avatar"
+        android:layout_toRightOf="@+id/chat_view_avatar"
+        android:layout_alignTop="@+id/chat_view_avatar"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:paddingBottom="4dp"
@@ -28,9 +28,9 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/message_body"
-        android:layout_below="@+id/name"
-        android:layout_alignLeft="@+id/name"
+        android:id="@+id/chat_textView_messageBody"
+        android:layout_below="@+id/chat_textView_name"
+        android:layout_alignLeft="@+id/chat_textView_name"
         android:background="@drawable/their_message"
         android:paddingVertical="12dp"
         android:paddingHorizontal="16dp"
