+++ app/src/test/java/com/github/gogetters/letsgo/game/PointTest.kt
